requirements.txt
langchain>=0.1.0
langchain-community>=0.0.10
langchain-core>=0.1.10

faiss-cpu>=1.7.4
sentence-transformers>=2.2.2
pypdf>=3.17.0
streamlit>=1.30.0
langchain-groq>=0.1.0
python-dotenv>=1.0.0
